pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["oci_vault_resolver.py"]

[project]
name = "oci-vault-mcp-resolver"
version = "1.0.0"
description = "Resolve OCI Vault secrets in Docker MCP Gateway configuration"
readme = "README.md"
license = {text = "AGPL-3.0"}
authors = [
    {name = "OCI Vault MCP Resolver Contributors"}
]
keywords = ["oci", "vault", "secrets", "mcp", "docker", "oracle-cloud"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "oci>=2.126.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",

    # Type checking
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.12",

    # Linting & formatting
    "flake8>=7.0.0",
    "black>=24.1.0",
    "pylint>=3.0.0",

    # Security
    "bandit>=1.7.7",

    # Pre-commit
    "pre-commit>=3.6.0",
]

[project.scripts]
oci-vault-resolve = "oci_vault_resolver:main"

[project.urls]
Homepage = "https://github.com/yourusername/oci-vault-mcp-resolver"
Documentation = "https://github.com/yourusername/oci-vault-mcp-resolver/blob/main/README.md"
Repository = "https://github.com/yourusername/oci-vault-mcp-resolver"
Issues = "https://github.com/yourusername/oci-vault-mcp-resolver/issues"

# ============================================================================
# Tool Configuration
# ============================================================================

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
    "-ra",                    # Show summary of all test outcomes
    "-q",                     # Quiet mode
    "--strict-markers",       # Raise error on unknown markers
    "--strict-config",        # Raise error on config issues
    "--cov=oci_vault_resolver", # Coverage for main module
    "--cov-report=term-missing", # Show missing lines in terminal
    "--cov-report=html",      # Generate HTML coverage report
    "--cov-report=xml",       # Generate XML for CI tools
]
markers = [
    "integration: Integration tests requiring real OCI credentials",
    "slow: Tests that take a long time to run",
]

[tool.coverage.run]
source = ["."]
branch = true
omit = [
    "tests/*",
    "setup.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "oci.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.pylint.main]
max-line-length = 100
disable = [
    "C0114",  # missing-module-docstring (we have it)
    "C0116",  # missing-function-docstring (some internal methods)
    "R0913",  # too-many-arguments (VaultResolver.__init__ needs them)
]

[tool.pylint.basic]
good-names = ["i", "j", "k", "e", "f", "id", "ttl"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B404", "B603"]  # Allow subprocess imports (needed for OCI CLI)

[tool.bandit.assert_used]
skips = ["*/test_*.py"]
